+++ tests/Unit/Support/VersionTest.php
@@ -40,6 +40,7 @@ class VersionTest extends AbstractUnit
         $this->assertEquals($this->bootstrap->config->path('phalcon.version'), $phalconVersion->get());
         
         // Test version->get()
+        $this->assertSame('2.6.0', $version->get());
         $this->assertNotEmpty($version->get());
         $this->assertIsString($version->get());